feat: add support flow state step to flow councils checkout
