+++ src/app/gooddollar/components/DistributionPoolFunding.tsx
@@ -22,12 +22,14 @@ import Offcanvas from "react-bootstrap/Offcanvas";
 import Accordion from "react-bootstrap/Accordion";
 import Stack from "react-bootstrap/Stack";
 import Image from "react-bootstrap/Image";
-import { Step } from "@/app/flow-councils/types/distributionPoolFunding";
-import EditStream from "./distribution-pool-funding/EditStream";
-import TopUp from "./distribution-pool-funding/TopUp";
-import Wrap from "./distribution-pool-funding/Wrap";
-import Review from "./distribution-pool-funding/Review";
-import Success from "./distribution-pool-funding/Success";
+import { Step } from "@/types/checkout";
+import EditStream from "@/components/checkout/EditStream";
+import TopUp from "@/components/checkout/TopUp";
+import Wrap from "@/components/checkout/Wrap";
+import SupportFlowState from "@/components/checkout/SupportFlowState";
+import Review from "@/components/checkout/Review";
+import Success from "@/components/checkout/Success";
+import { SupEvent } from "@/app/api/flow-council/db";
 import { Network } from "@/types/network";
 import DistributionPoolDetails from "./DistributionPoolDetails";
 import useFlowingAmount from "@/hooks/flowingAmount";
@@ -43,7 +45,14 @@ import {
   formatNumberWithCommas,
   roundWeiAmount,
 } from "@/lib/utils";
-import { SECONDS_IN_MONTH, MAX_FLOW_RATE, ZERO_ADDRESS } from "@/lib/constants";
+import {
+  SECONDS_IN_MONTH,
+  MAX_FLOW_RATE,
+  ZERO_ADDRESS,
+  FLOW_STATE_RECEIVER,
+} from "@/lib/constants";
+import { getSupportFlowStateConfig } from "@/lib/supportFlowStateConfig";
+import { getSocialShare } from "../lib/socialShare";
 
 const SF_ACCOUNT_QUERY = gql`
   query SFAccountQuery($userAddress: String!, $token: String!) {
@@ -95,10 +104,9 @@ export default function DistributionPoolFunding(props: {
 
   const [step, setStep] = useState<Step>(Step.SELECT_AMOUNT);
   const [amountPerTimeInterval, setAmountPerTimeInterval] = useState("");
-  const [timeInterval, setTimeInterval] = useState<TimeInterval>(
-    TimeInterval.MONTH,
-  );
   const [newFlowRate, setNewFlowRate] = useState("");
+  const [newFlowRateToFlowState, setNewFlowRateToFlowState] = useState("");
+  const [supportFlowStateAmount, setSupportFlowStateAmount] = useState("");
   const [wrapAmount, setWrapAmount] = useState("");
   const [underlyingTokenAllowance, setUnderlyingTokenAllowance] = useState("0");
   const [sfFramework, setSfFramework] = useState<Framework | null>(null);
@@ -108,7 +116,7 @@ export default function DistributionPoolFunding(props: {
 
   const { isMobile } = useMediaQuery();
   const { address } = useAccount();
-  const { council, councilMetadata, token, gdaPool } = useCouncil();
+  const { council, token, gdaPool } = useCouncil();
   const {
     areTransactionsLoading,
     completedTransactions,
@@ -199,6 +207,14 @@ export default function DistributionPoolFunding(props: {
     superTokenBalance > BigInt(0)
       ? true
       : false;
+  const socialShare = getSocialShare({
+    councilUiLink: `https://flowstate.network/gooddollar`,
+  });
+  const supportFlowStateConfig = getSupportFlowStateConfig(token.symbol);
+  const flowRateToFlowState = superfluidQueryRes?.account?.outflows?.find(
+    (outflow: { receiver: { id: string } }) =>
+      outflow.receiver.id === FLOW_STATE_RECEIVER,
+  );
 
   const flowRateToReceiver = useMemo(() => {
     if (address && gdaPool) {
@@ -241,11 +257,11 @@ export default function DistributionPoolFunding(props: {
   }, [poolMemberships]);
 
   const calcLiquidationEstimate = useCallback(
-    (amountPerTimeInterval: string, timeInterval: TimeInterval) => {
+    (amountPerTimeInterval: string) => {
       if (address) {
         const newFlowRate =
           parseEther(amountPerTimeInterval.replace(/,/g, "")) /
-          BigInt(fromTimeUnitsToSeconds(1, unitOfTime[timeInterval]));
+          BigInt(fromTimeUnitsToSeconds(1, unitOfTime[TimeInterval.MONTH]));
         const accountFlowRate =
           userAccountSnapshot?.totalNetFlowRate ?? "0" + membershipsInflowRate;
 
@@ -291,8 +307,8 @@ export default function DistributionPoolFunding(props: {
   );
 
   const liquidationEstimate = useMemo(
-    () => calcLiquidationEstimate(amountPerTimeInterval, timeInterval),
-    [calcLiquidationEstimate, amountPerTimeInterval, timeInterval],
+    () => calcLiquidationEstimate(amountPerTimeInterval),
+    [calcLiquidationEstimate, amountPerTimeInterval],
   );
 
   const transactions = useMemo(() => {
@@ -391,19 +407,19 @@ export default function DistributionPoolFunding(props: {
 
       setAmountPerTimeInterval(formatNumberWithCommas(currentStreamValue));
     })();
-  }, [address, flowRateToReceiver, timeInterval]);
+  }, [address, flowRateToReceiver]);
 
   useEffect(() => {
     if (!areTransactionsLoading && amountPerTimeInterval) {
       const newFlowRate =
         parseEther(amountPerTimeInterval.replace(/,/g, "")) /
-        BigInt(fromTimeUnitsToSeconds(1, unitOfTime[timeInterval]));
+        BigInt(fromTimeUnitsToSeconds(1, unitOfTime[TimeInterval.MONTH]));
 
       if (newFlowRate < MAX_FLOW_RATE) {
         setNewFlowRate(newFlowRate.toString());
       }
     }
-  }, [areTransactionsLoading, amountPerTimeInterval, timeInterval]);
+  }, [areTransactionsLoading, amountPerTimeInterval]);
 
   useEffect(() => {
     (async () => {
@@ -430,9 +446,46 @@ export default function DistributionPoolFunding(props: {
     })();
   }, [address, ethersProvider, distributionTokenAddress, network]);
 
+  useEffect(() => {
+    (async () => {
+      if (step !== Step.SUPPORT) {
+        return;
+      }
+
+      const currentStreamValue = roundWeiAmount(
+        BigInt(flowRateToFlowState?.currentFlowRate ?? 0) *
+          BigInt(SECONDS_IN_MONTH),
+        4,
+      );
+
+      setSupportFlowStateAmount(
+        formatNumberWithCommas(
+          `${
+            Number(currentStreamValue) +
+            supportFlowStateConfig.suggestedFlowStateDonation
+          }`,
+        ),
+      );
+    })();
+  }, [address, flowRateToFlowState, supportFlowStateConfig, step]);
+
+  useEffect(() => {
+    if (areTransactionsLoading) {
+      return;
+    }
+
+    setNewFlowRateToFlowState(
+      supportFlowStateAmount
+        ? (
+            parseEther(supportFlowStateAmount.replace(/,/g, "")) /
+            BigInt(fromTimeUnitsToSeconds(1, unitOfTime[TimeInterval.MONTH]))
+          ).toString()
+        : "",
+    );
+  }, [areTransactionsLoading, supportFlowStateAmount]);
+
   const updateWrapAmount = (
     amountPerTimeInterval: string,
-    timeInterval: TimeInterval,
     liquidationEstimate: number | null,
   ) => {
     if (amountPerTimeInterval) {
@@ -482,12 +535,23 @@ export default function DistributionPoolFunding(props: {
 
       const newFlowRate =
         parseEther(amountPerTimeInterval.replace(/,/g, "")) /
-        BigInt(fromTimeUnitsToSeconds(1, unitOfTime[timeInterval]));
+        BigInt(fromTimeUnitsToSeconds(1, unitOfTime[TimeInterval.MONTH]));
 
       if (newFlowRate < MAX_FLOW_RATE) {
         setNewFlowRate(newFlowRate.toString());
       }
     }
+  };
+
+  const sendSupEvent = (event: SupEvent) => {
+    fetch("/api/good-dollar/sup", {
+      method: "POST",
+      body: JSON.stringify({
+        address,
+        chainId: network.id,
+        event,
+      }),
+    });
   };
 
   return (
@@ -509,6 +573,7 @@ export default function DistributionPoolFunding(props: {
           (Number(gdaPool.totalUnits) > 0 || BigInt(flowRateToReceiver) > 0) ? (
             <Accordion activeKey={step} className="mt-4">
               <EditStream
+                isFundingDistributionPool={true}
                 isSelected={step === Step.SELECT_AMOUNT}
                 setStep={(step) => setStep(step)}
                 token={token}
@@ -519,7 +584,10 @@ export default function DistributionPoolFunding(props: {
                   const newAmount =
                     parseEther(amount.replace(/,/g, "")) /
                       BigInt(
-                        fromTimeUnitsToSeconds(1, unitOfTime[timeInterval]),
+                        fromTimeUnitsToSeconds(
+                          1,
+                          unitOfTime[TimeInterval.MONTH],
+                        ),
                       ) <
                     MAX_FLOW_RATE
                       ? amount
@@ -528,31 +596,20 @@ export default function DistributionPoolFunding(props: {
                   setAmountPerTimeInterval(newAmount);
                   updateWrapAmount(
                     newAmount,
-                    timeInterval,
-                    calcLiquidationEstimate(newAmount, timeInterval),
+                    calcLiquidationEstimate(newAmount),
                   );
                 }}
                 newFlowRate={newFlowRate}
                 wrapAmount={wrapAmount}
-                timeInterval={timeInterval}
-                setTimeInterval={(timeInterval) => {
-                  setTimeInterval(timeInterval);
-                  updateWrapAmount(
-                    amountPerTimeInterval,
-                    timeInterval,
-                    calcLiquidationEstimate(
-                      amountPerTimeInterval,
-                      timeInterval,
-                    ),
-                  );
-                }}
                 superTokenBalance={superTokenBalance}
                 isSuperTokenPure={isSuperTokenPure}
                 hasSufficientBalance={
                   !!hasSufficientEthBalance && !!hasSuggestedTokenBalance
                 }
+                docsLink="https://docs.flowstate.network/flow-councils/grow-the-pie"
               />
               <TopUp
+                isFundingDistributionPool={true}
                 step={step}
                 setStep={(step) => setStep(step)}
                 newFlowRate={newFlowRate}
@@ -571,6 +628,7 @@ export default function DistributionPoolFunding(props: {
               />
               {!isSuperTokenPure && (
                 <Wrap
+                  isFundingDistributionPool={true}
                   step={step}
                   setStep={setStep}
                   wrapAmount={wrapAmount}
@@ -578,9 +636,25 @@ export default function DistributionPoolFunding(props: {
                   token={token}
                   superTokenBalance={superTokenBalance}
                   underlyingTokenBalance={underlyingTokenBalance}
+                  newFlowRate={newFlowRate}
                 />
               )}
+              <SupportFlowState
+                isFundingDistributionPool={true}
+                network={network}
+                token={token}
+                step={step}
+                setStep={(step) => setStep(step)}
+                supportFlowStateAmount={supportFlowStateAmount}
+                setSupportFlowStateAmount={setSupportFlowStateAmount}
+                newFlowRateToFlowState={newFlowRateToFlowState}
+                flowRateToFlowState={
+                  flowRateToFlowState?.currentFlowRate ?? "0"
+                }
+                isSuperTokenPure={isSuperTokenPure}
+              />
               <Review
+                isFundingDistributionPool={true}
                 step={step}
                 setStep={(step) => setStep(step)}
                 network={network}
@@ -597,21 +671,20 @@ export default function DistributionPoolFunding(props: {
                 amountPerTimeInterval={amountPerTimeInterval}
                 newFlowRate={newFlowRate}
                 wrapAmount={wrapAmount}
-                newFlowRateToFlowState={"0"}
-                flowRateToFlowState={"0"}
-                timeInterval={timeInterval}
-                supportFlowStateAmount={"0"}
-                supportFlowStateTimeInterval={TimeInterval.MONTH}
+                newFlowRateToFlowState={newFlowRateToFlowState}
+                flowRateToFlowState={
+                  flowRateToFlowState?.currentFlowRate ?? "0"
+                }
+                supportFlowStateAmount={supportFlowStateAmount}
                 isSuperTokenPure={isSuperTokenPure}
                 superTokenBalance={superTokenBalance}
                 underlyingTokenBalance={underlyingTokenBalance}
               />
               <Success
                 step={step}
-                councilName={councilMetadata.name}
-                chainId={network.id}
-                councilUiLink={`https://flowstate.network/gooddollar`}
                 newFlowRate={newFlowRate}
+                socialShare={socialShare}
+                onClick={() => sendSupEvent("shared-distribution")}
               />
             </Accordion>
           ) : (
